text-detector/requirements.txt
confluent-kafka==2.10.0
python-dotenv==1.1.0
